Cleanup: moved a few functions into separate files
