+++ stm32/dodekaedri.c
@@ -1,8 +1,11 @@
-#include "stm32f4xx.h"
-#include "stm32f4xx_rcc.h"
-#include "stm32f4xx_gpio.h"
-#include "stm32f4xx_i2c.h"
-#include "stm32f4xx_usart.h"
+#include <stm32f4xx.h>
+#include <stm32f4xx_rcc.h>
+#include <stm32f4xx_gpio.h>
+#include <stm32f4xx_i2c.h>
+#include <stm32f4xx_usart.h>
+
+#include "uart.h"
+#include "i2c.h"
 
 // SI5351 register map generated by ClockBuilder
 // (only for initial test, probably not going to be actually used)
@@ -16,88 +19,12 @@ void SystemInit() {
 const uint8_t addr_synth = 0x60; // SI5351 synthesizer I2C address
 const uint8_t addr_rfc = 0x1A;   // WM8731 used for zero-IF signal
 const uint8_t addr_afc = 0x1B;   // WM8731 used for microphone and headphone audio
-
-int i2cdelay = 20;
-void i2c_delay() {
-	int d;
-	for(d = 0; d < i2cdelay; d++) __asm("nop");
-}
-
-// SCL 8, SDA 9
-#define I2C_SCL 0x100
-#define I2C_SDA 0x200
-
-void writereg(uint8_t addr, uint8_t reg, uint8_t v) {
-	int i;
-	uint32_t odr = GPIOB->ODR & (~(I2C_SCL | I2C_SDA));
-	uint32_t allbits;
-	// start
-	GPIOB->ODR = odr | I2C_SCL;
-	i2c_delay();
-	GPIOB->ODR = odr;
-	i2c_delay();
-
-	allbits =
-	(addr << (9*2+2)) | (0 << (9*2+1)) | (1 << (9*2)) |
-	(reg  << (9*1+1)) |                  (1 << (9*1)) |
-	(v    << (9*0+1)) |                  (1 << (9*0));
-	for(i = 0; i < 9*3; i++) {
-		uint32_t odr_bit;
-		if(allbits & (1 << (9*3-1)))
-			odr_bit = odr | I2C_SDA;
-		else
-			odr_bit = odr;
-		allbits <<= 1;
-		// send bit
-		GPIOB->ODR = odr_bit;
-		i2c_delay();
-		GPIOB->ODR = odr_bit | I2C_SCL;
-		i2c_delay();
-		GPIOB->ODR = odr_bit;
-		i2c_delay();
-	}
-
-	// stop
-	GPIOB->ODR = odr | I2C_SCL;
-	i2c_delay();
-	GPIOB->ODR = odr | I2C_SCL | I2C_SDA;
-	i2c_delay();
-}
-
-
-int usart_enabled = 0;
-void print(char *data) {
-	/* Send text to USART1 */
-	if(usart_enabled == 0) return;
-	for(; *data != '\0'; data++) {
-		/*while(!USART_GetFlagStatus(USART1, USART_FLAG_TXE));
-		USART_SendData(USART1, (uint8_t)*data);*/
-		while(!(USART1->SR & USART_SR_TXE));
-		USART1->DR = (uint8_t)*data;
-		//ITM_SendChar(*data);
-	}
-}
-
-const char hexnumbers[16] = "0123456789ABCDEF";
-void printhex(uint32_t v) {
-	char text[9];
-	int i;
-	for(i = 0; i < 8; i++)
-		text[i] = hexnumbers[0xF & v >> (4*(7-i))];
-	text[8] = '\0';
-	print(text);
-}
-
 void assert_failed(uint8_t *file, uint32_t line) {
-	char l[9] = "      \r\n";
 	print("\r\nAssert failed: ");
 	print((char*)file);
-	l[5] = '0' + line % 10;
-	l[4] = '0' + (line / 10) % 10;
-	l[3] = '0' + (line / 100) % 10;
-	l[2] = '0' + (line / 1000) % 10;
-	l[1] = '0' + (line / 10000);
-	print(l);
+	print(" ");
+	printdec(line);
+	print("\r\n");
 }
 
 // The WM8731 registers have 7-bit addresses and 9-bit data
@@ -167,14 +94,9 @@ int main() {
 	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOB, ENABLE);
 	RCC_AHB1PeriphClockCmd(RCC_AHB1Periph_GPIOC, ENABLE);
 
-	// bit banged I2C pins
-	GPIO_Init(GPIOB, &(GPIO_InitTypeDef) {
-		.GPIO_Pin = GPIO_Pin_8 | GPIO_Pin_9,
-		.GPIO_Mode = GPIO_Mode_OUT,
-		.GPIO_Speed = GPIO_Speed_50MHz,
-		.GPIO_OType = GPIO_OType_OD,
-		.GPIO_PuPd = GPIO_PuPd_UP
-	});
+	uart_init();
+
+	i2c_init();
 
 	// init SI5351
 	int reg_i;
@@ -186,33 +108,6 @@ int main() {
 	for(reg_i = 0; reg_i < RFC_INIT_SIZE; reg_i++) {
 		writereg(addr_rfc, rfc_init[reg_i][0], rfc_init[reg_i][1]);
 	}
-
-
-	// USART1 initialization
-	// External USB-UART converter is connected to PA9
-	// because the built-in converter on Nucleo board in USART2
-	// seems to work poorly.
-	RCC_APB2PeriphClockCmd(RCC_APB2Periph_USART1, ENABLE);
-	GPIO_PinAFConfig(GPIOA, GPIO_PinSource9, GPIO_AF_USART1); // USART1_TX
-	GPIO_PinAFConfig(GPIOA, GPIO_PinSource10, GPIO_AF_USART1); // USART1_RX
-	GPIO_Init(GPIOA, &(GPIO_InitTypeDef) {
-		.GPIO_Pin = GPIO_Pin_9 | GPIO_Pin_10,
-		.GPIO_Mode = GPIO_Mode_AF,
-		.GPIO_Speed = GPIO_Speed_50MHz,
-		.GPIO_OType = GPIO_OType_PP,
-		.GPIO_PuPd = GPIO_PuPd_UP
-	});
-	USART_Init(USART1, &(USART_InitTypeDef) {
-		.USART_BaudRate = 115200,
-		.USART_WordLength = USART_WordLength_8b,
-		.USART_StopBits = USART_StopBits_1,
-		.USART_Parity = USART_Parity_No,
-		.USART_Mode = USART_Mode_Tx | USART_Mode_Rx,
-		.USART_HardwareFlowControl = USART_HardwareFlowControl_None
-	});
-	USART_Cmd(USART1, ENABLE);
-	usart_enabled = 1;
-	print("\r\nHello world\r\n");
 
 
 	// SPI1 initialization
@@ -249,7 +144,8 @@ int main() {
 	});
 	SPI_Cmd(SPI1, ENABLE);
 
-	// I2S PLL initialization
+
+	// Main PLL and I2S PLL initialization
 	RCC_PLLI2SCmd(DISABLE); // PLLI2SConfig should be called with PLLI2S disabled
 
 	// PLLI2S clock source is common with main PLL configured with RCC_PLLConfig.
