+++ stm32/i2c.h
@@ -1,0 +1,2 @@
+void writereg(uint8_t, uint8_t, uint8_t);
+void i2c_init();
