+++ stm32/uart.h
@@ -1,0 +1,4 @@
+void print(char *);
+void printhex(uint32_t);
+void printdec(uint32_t);
+void uart_init();
